requirements.txt
django-tinymce
six
django_apscheduler
django-filter
pytils
